Add FRESH-TEST, TEST-SYSTEM, more
Fix the return value of RUN-TESTS to actually match CL-TEST-GRID.

FRESH-TEST fires up a remote lisp and runs the tests in it.

TEST-SYSTEM calls ASDF:TEST-SYSTEM but returns the value of the last
call to RUN-TESTS. If :quit is nonnil, it causes the lisp image to
quit and return a status code indicating the success of the tests.

Add docstrings file.